unit test 코드 수정

+++ src/test/java/com/knowget/knowgetbackend/global/entity/SuccessCaseTest.java
@@ -88,7 +88,7 @@ class SuccessCaseTest {
 	@DisplayName("SuccessCase 엔티티 승인 상태 업데이트 테스트")
 	public void testUpdateIsApproved() {
 		// Given
-		Integer newIsApproved = 1;
+		Short newIsApproved = 1;
 
 		// When
 		successCase.updateIsApproved(newIsApproved);
